test: Move Foo.java and ToEvaluate.java to testclasses (#2164)
